requirements.txt
adium-theme-ubuntu==0.3.4
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==2.1.3
boto==2.48.0
bottle==0.12.9
bs4==0.0.1
celery==3.1.25
configparser==3.5.0
cssmin==0.2.0
decorator==4.0.6
Django==1.8.18
django-assets==0.11
django-celery==3.2.1
django-debug-toolbar==1.6
django-ses==0.7.0
django-storages==1.1.8
djangorestframework==3.8.2
ecdsa==0.13
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==2.1.6
gunicorn==0.17.4
hamlpy==0.82.2
html5lib==1.0.1
httpretty==0.8.0
ipython==2.4.1
ipython-genutils==0.2.0
Jinja2==2.8
jsmin==2.2.1
jsonschema==2.6.0
jupyter-core==4.4.0
kombu==3.0.37
libsass==0.13.3
Markdown==2.4
MarkupSafe==0.23
mistune==0.8.3
mock==1.3.0
mysql-connector-python==2.0.4
mysql-utilities==1.6.1
nbconvert==5.3.1
nbformat==4.4.0
pandocfilters==1.4.2
paramiko==1.16.0
pathlib2==2.3.2
pbr==3.1.1
pexpect==4.0.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5
pycrypto==2.6.1
pycurl==7.43.0
Pygments==1.6
pygobject==3.20.0
PyMySQL==0.7.2
pyodbc==3.0.7
pysqlite==2.7.0
pytz==2015.2
pyzmq==15.2.0
scandir==1.7
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.4
testpath==0.3.1
tornado==4.2.1
traitlets==4.3.2
unity-lens-photos==1.0
urllib3==1.15.1
vboxapi==1.0
virtualenv==15.1.0
wcwidth==0.1.7
webassets==0.11
webencodings==0.5.1
